requirements.txt
asgiref==3.8.1
Django==5.1.6
django-cors-headers==4.7.0
django-filter==25.1
django-js-asset==3.1.0
django-mptt==0.16.0
djangorestframework==3.15.2
drf-yasg==1.21.9
inflection==0.5.1
packaging==24.2
pillow==11.1.0
psycopg2==2.9.10
pytz==2025.1
PyYAML==6.0.2
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1
